♻️factorisation of code to be shared with NewAccessProfileCommand
