+++ src/validator/requiredValidator.ts
@@ -1,0 +1,5 @@
+import { Validator } from "./validator";
+
+export const requiredValidator = new Validator({
+    required: true
+});
